fix require cycle closes #7
