+++ src/modules/DecksList/index.ts
@@ -1,4 +1,3 @@
 import DecksList from './components/DecksList';
-import useDecks from './hooks/useDecks';
 
-export { DecksList, useDecks };
+export { DecksList };
